refactor nft module
