+++ esdt-system-sc-mock/Cargo.toml
@@ -9,7 +9,7 @@ publish = false
 path = "src/esdt_system_sc_mock.rs"
 
 [dependencies.elrond-wasm]
-version = "0.31.0"
+version = "0.31.1"
 
 [dev-dependencies.elrond-wasm-debug]
-version = "0.31.0"
+version = "0.31.1"
