+++ esdt-system-sc-mock/meta/Cargo.toml
@@ -9,4 +9,4 @@ publish = false
 path = ".."
 
 [dependencies.elrond-wasm-debug]
-version = "0.31.0"
+version = "0.31.1"
