+++ esdt-system-sc-mock/wasm/Cargo.toml
@@ -19,10 +19,10 @@ panic = "abort"
 path = ".."
 
 [dependencies.elrond-wasm-node]
-version = "0.31.0"
+version = "0.31.1"
 
 [dependencies.elrond-wasm-output]
-version = "0.31.0"
+version = "0.31.1"
 features = ["wasm-output-mode"]
 
 [workspace]
