+++ nft-minter/Cargo.toml
@@ -9,13 +9,13 @@ publish = false
 path = "src/lib.rs"
 
 [dependencies.elrond-wasm]
-version = "0.31.0"
+version = "0.31.1"
 
 [dev-dependencies.esdt-system-sc-mock]
 path = "../esdt-system-sc-mock"
 
 [dev-dependencies.elrond-wasm-debug]
-version = "0.31.0"
+version = "0.31.1"
 
 [dev-dependencies]
 num-bigint = "0.4.2"
