+++ nft-minter/meta/Cargo.toml
@@ -11,4 +11,4 @@ authors = [ "Dorin Marian Iancu, dorin.iancu@elrond.com" ]
 path = ".."
 
 [dependencies.elrond-wasm-debug]
-version = "0.31.0"
+version = "0.31.1"
