+++ nft-minter/src/lib.rs
@@ -3,23 +3,27 @@
 elrond_wasm::imports!();
 
 pub mod admin_whitelist;
+pub mod brand_creation;
 pub mod common_storage;
 pub mod nft_attributes_builder;
 pub mod nft_marketplace_interactor;
-pub mod nft_module;
+pub mod nft_minting;
 pub mod nft_tier;
 pub mod royalties;
 pub mod unique_id_mapper;
+pub mod views;
 
 #[elrond_wasm::contract]
 pub trait NftMinter:
     common_storage::CommonStorageModule
     + admin_whitelist::AdminWhitelistModule
-    + nft_module::NftModule
+    + brand_creation::BrandCreationModule
+    + nft_minting::NftMintingModule
     + nft_tier::NftTierModule
     + nft_attributes_builder::NftAttributesBuilderModule
     + royalties::RoyaltiesModule
     + nft_marketplace_interactor::NftMarketplaceInteractorModule
+    + views::ViewsModule
 {
     #[init]
     fn init(
