+++ nft-minter/wasm/Cargo.toml
@@ -24,8 +24,8 @@ panic = "abort"
 path = ".."
 
 [dependencies.elrond-wasm-node]
-version = "0.31.0"
+version = "0.31.1"
 
 [dependencies.elrond-wasm-output]
-version = "0.31.0"
+version = "0.31.1"
 features = [ "wasm-output-mode",]
