+++ royalties-handler/Cargo.toml
@@ -12,13 +12,13 @@ path = "src/lib.rs"
 path = "../nft-minter"
 
 [dependencies.elrond-wasm]
-version = "0.31.0"
+version = "0.31.1"
 
 [dev-dependencies.esdt-system-sc-mock]
 path = "../esdt-system-sc-mock"
 
 [dev-dependencies.elrond-wasm-debug]
-version = "0.31.0"
+version = "0.31.1"
 
 [dev-dependencies]
 num-bigint = "0.4.2"
